requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
python-dateutil>=2.8.2
schedule>=1.2.0
loguru>=0.7.2
tenacity>=8.2.3
flask>=3.0.0
flask-cors>=4.0.0
polymarket-apis>=0.5.0
numpy>=1.26.0
